Refactoring: исправление архитектурных ошибок
