+++ src/entities/ResultPanel/ui/ResultPanel.module.css
@@ -1,4 +1,4 @@
-@import url('../../styles/mixins.css');
+@import url('../../../shared/styles/mixins.css');
 
 .result-section {
     @mixin responsive width, 587;
