+++ src/features/main/MainContent/index.js
@@ -1,0 +1,1 @@
+export { MainContent } from './ui/MainContent'
