+++ src/features/main/search/ui/Search.module.css
@@ -1,4 +1,4 @@
-@import url('../../styles/mixins.css');
+@import url('../../../shared/styles/mixins.css');
 
 .search-form {
     /*@mixin responsive margin-top, 200;*/
