+++ src/pages/mainPage/ui/MainPage.jsx
@@ -1,0 +1,15 @@
+import { Content } from "@/widgets/Content";
+import { Header } from "@/widgets/Header";
+import { Footer } from "@/widgets/Footer";
+import { Background } from "@/widgets/Background";
+
+export const MainPage = () => {
+  return (
+    <>
+      <Header />
+      <Background />
+      <Content />
+      <Footer />
+    </>
+  )
+}
