--- src/pages/mainPage/ui/Page.jsx
@@ -1,9 +1,0 @@
-import { MainContent } from "@/features/main/MainConent/ui/MainContent"
-
-export const MainPage = () => {
-  return (
-    <>
-      <MainContent />
-    </>
-  )
-}
