+++ src/widgets/Content/index.js
@@ -1,1 +1,1 @@
-export { Content } from './ui/content';
+export { Content } from './ui/Сontent';
