+++ src/widgets/Content/ui/Сontent.jsx
@@ -1,10 +1,11 @@
 import styles from './Сontent.module.css';
+import { MainContent } from '@/features/main/MainContent';
 
-export function Content() {
+export const Content = () => {
   return (
     <main className={styles.main}>
       <div className={styles.container}>
-
+        <MainContent />
       </div>
     </main>
   )
