+++ src/widgets/Content/ui/Сontent.module.css
@@ -1,4 +1,4 @@
-@import url('../../styles/mixins.css');
+@import url('../../../shared/styles/mixins.css');
 
 
 .container {
