+++ src/widgets/background/ui/Background.module.css
@@ -1,4 +1,4 @@
-@import url('../../styles/mixins.css');
+@import url('../../../shared/styles/mixins.css');
 
 .background {
     position: absolute;
