+++ src/widgets/header/ui/Header.jsx
@@ -1,7 +1,7 @@
 import styles from './Header.module.css';
 import multiLogo from '../../../shared/assets/images/multiLogo.png';
 
-export function Header() {
+export const Header = () => {
   return (
     <header className={styles.header}>
       <img className={styles.logo} src={multiLogo} alt="Логотип Multigrab" draggable="false" />
